Update sdkVersion to 35.0.0 Declare PATH into Dockerfile Add --rebuild-ecosystem arg to compose script Expose ports 5555 and 5554 to improve debugging Use Nexus 5 as device

+++ ecosystem/compose.sh
@@ -2,7 +2,11 @@
 
 docker-compose down
 yes | docker system prune
-docker image rm -f ecosystem_expo:latest
+
+if [ "$1" == "--rebuild-ecosystem" ]; then
+  docker image rm -f ecosystem_expo:latest
+fi
+
 (cd ../code; npm i; npm audit fix)
 docker-compose up
